Update Readme.md
Line 125 - there is no such thing as a "code pen", thus adjusted to reflect proper use.

Line 166 - Promise and fetch - Fetch is an API and Promise is an abstraction, or more precisely a constructor, but should be capitalized, as they are both APIs as well. And for consistency. Fetch is commonly capitalized across the web, again and again. For instance see Chris Coyier's CSS Tricks, for instance at: [Using Fetch](https://css-tricks.com/using-fetch/)