"its", not "it's"
"it's" is a contraction of "it is". "its" is the possessive: "belonging to it". Thus sentences such as "Cedar uses amCharts library as its charting engine" need to use "its", not "it's". Of all the orthographical battles, this is one that prescriptive grammarians are steadily losing, but I try to fix this error where I can.